+++ delta_backend/tests/test_convert_dose_sequence.py
@@ -2,7 +2,7 @@ import copy
 import json
 import unittest
 from tests.utils_for_converter_tests import ValuesForTests
-from Converter import Converter
+from delta_converter import Converter
 
 
 class TestDoseSequenceToFlatJson(unittest.TestCase):
@@ -13,7 +13,7 @@ class TestDoseSequenceToFlatJson(unittest.TestCase):
     def _run_test(self, expected_result):
         """Helper function to run the test"""
         self.converter = Converter(json.dumps(self.request_json_data))
-        flat_json = self.converter.runConversion(self.request_json_data, False, True)
+        flat_json = self.converter.runConversion()
         self.assertEqual(flat_json["DOSE_SEQUENCE"], expected_result)
         
     def test_dose_sequence_present_int(self):
