+++ delta_backend/tests/test_convert_person_forename.py
@@ -2,7 +2,7 @@ import copy
 import json
 import unittest
 from tests.utils_for_converter_tests import ValuesForTests
-from Converter import Converter
+from delta_converter import Converter
 
 
 class TestPersonForeNameToFlatJson(unittest.TestCase):
@@ -107,5 +107,5 @@ class TestPersonForeNameToFlatJson(unittest.TestCase):
     def _run_test(self, expected_forename):
         """Helper function to run the test"""
         self.converter = Converter(json.dumps(self.request_json_data))
-        flat_json = self.converter.runConversion(self.request_json_data, False, True)
+        flat_json = self.converter.runConversion()
         self.assertEqual(flat_json["PERSON_FORENAME"], expected_forename)
