+++ delta_backend/tests/test_convert_practitioner_forename.py
@@ -2,7 +2,7 @@ import copy
 import json
 import unittest
 from tests.utils_for_converter_tests import ValuesForTests
-from Converter import Converter
+from delta_converter import Converter
 
 
 class TestPractitionerForeNameToFlatJson(unittest.TestCase):
@@ -121,6 +121,6 @@ class TestPractitionerForeNameToFlatJson(unittest.TestCase):
     def _run_practitioner_test(self, expected_forename):
         """Helper function to run the test"""
         self.converter = Converter(json.dumps(self.request_json_data))
-        flat_json = self.converter.runConversion(self.request_json_data, False, True)
+        flat_json = self.converter.runConversion()
         self.assertEqual(flat_json["PERFORMING_PROFESSIONAL_FORENAME"], expected_forename)
