+++ delta_backend/tests/test_convert_site_code.py
@@ -2,7 +2,7 @@ import copy
 import json
 import unittest
 from tests.utils_for_converter_tests import ValuesForTests
-from Converter import Converter
+from delta_converter import Converter
 
 
 class TestPersonSiteCodeToFlatJson(unittest.TestCase):
@@ -124,5 +124,5 @@ class TestPersonSiteCodeToFlatJson(unittest.TestCase):
     def _run_site_code_test(self, expected_site_code):
         """Helper function to run the test"""
         self.converter = Converter(json.dumps(self.request_json_data))
-        flat_json = self.converter.runConversion(self.request_json_data, False, True)
+        flat_json = self.converter.runConversion()
         self.assertEqual(flat_json.get("SITE_CODE"), expected_site_code)
