+++ delta_backend/tests/test_convert_site_uri.py
@@ -2,7 +2,7 @@ import copy
 import json
 import unittest
 from tests.utils_for_converter_tests import ValuesForTests
-from Converter import Converter
+from delta_converter import Converter
 
 
 class TestPersonSiteUriToFlatJson(unittest.TestCase):
@@ -96,6 +96,6 @@ class TestPersonSiteUriToFlatJson(unittest.TestCase):
     def _run_site_uri_test(self, expected_site_code):
         """Helper function to run the test"""
         self.converter = Converter(json.dumps(self.request_json_data))
-        flat_json = self.converter.runConversion(self.request_json_data, False, True)
+        flat_json = self.converter.runConversion()
         self.assertEqual(flat_json.get("SITE_CODE_TYPE_URI"), expected_site_code)
